Remove old files related to the Blaze web flow
